[Port from Java] Fix HTTP response processing for WWW-Authenticate headers
The original code was able to process only one WWW-Authenticate
header in an HTTP response, and if this header was not one of
two expected, authentication failed regardless of that there
could be other headers in the response.

All WWW-Authenticate headers in an HTTP response have to be
browsed to find one of supported, i.e. Basic or Digest.
By that if both are present, the Digest one should be used
as more preferable.

Bug 357719 - HTTP authentication fails when server returns multiple WWW-Authenticate headers

refs https://bugs.eclipse.org/bugs/show_bug.cgi?id=357719
https://git.eclipse.org/r/#/c/13285/